+++ packages/ubcli/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - `ubcli generateNginxCfg` - added `/metrics` endpoint restriction generation using
+   ubConfig `metrics.allowedFrom` setting. 
 
 ### Changed